+++ evals/context-query/fixture/go/main.go
@@ -1,0 +1,7 @@
+package main
+
+import "./util"
+
+func main() {
+	println(util.Total(2))
+}